ad closing small holes
